py/v1_uploadPictureDirectly/requirements.txt
Pillow>=10.0.0
pyserial>=3.5
